Refactor test-client -> integration-test
