+++ test/integration-test-common/meson.build
@@ -1,0 +1,2 @@
+integration_test_common_inc = include_directories('.')
+integration_test_common_srcs = files('integration-test-common.c')
