+++ test/integration-tests/test-adapts-to-screen-size.c
@@ -9,25 +9,47 @@
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-#include "test-client-common.h"
+#include "integration-test-common.h"
 
 static GtkWindow* window;
 
 static void callback_0()
 {
+    EXPECT_MESSAGE(zwlr_layer_surface_v1 .set_size 0 0);
+    EXPECT_MESSAGE(.create_buffer 1920 1080); // size must match DEFAULT_OUTPUT_WIDTH/DEFAULT_OUTPUT_HEIGHT in common.h
+
     window = create_default_window();
+
     gtk_layer_init_for_window(window);
-    gtk_layer_set_layer(window, GTK_LAYER_SHELL_LAYER_OVERLAY);
-    ASSERT_EQ(gtk_layer_get_layer(window), GTK_LAYER_SHELL_LAYER_OVERLAY, "%d");
-    gtk_widget_show_all(GTK_WIDGET(window));
-    gtk_layer_set_layer(window, GTK_LAYER_SHELL_LAYER_TOP);
-    ASSERT_EQ(gtk_layer_get_layer(window), GTK_LAYER_SHELL_LAYER_TOP, "%d");
-    gtk_layer_set_layer(window, GTK_LAYER_SHELL_LAYER_BACKGROUND);
-    ASSERT_EQ(gtk_layer_get_layer(window), GTK_LAYER_SHELL_LAYER_BACKGROUND, "%d");
     gtk_layer_set_layer(window, GTK_LAYER_SHELL_LAYER_BOTTOM);
-    ASSERT_EQ(gtk_layer_get_layer(window), GTK_LAYER_SHELL_LAYER_BOTTOM, "%d");
+    gtk_layer_set_namespace(window, "foobar");
+    gtk_layer_set_anchor(window, GTK_LAYER_SHELL_EDGE_BOTTOM, TRUE);
+    gtk_layer_set_anchor(window, GTK_LAYER_SHELL_EDGE_TOP, TRUE);
+    gtk_layer_set_anchor(window, GTK_LAYER_SHELL_EDGE_LEFT, TRUE);
+    gtk_layer_set_anchor(window, GTK_LAYER_SHELL_EDGE_RIGHT, TRUE);
+
+    gtk_widget_set_size_request(GTK_WIDGET(window), 600, 700);
+    gtk_window_resize(window, 1, 1);
+    gtk_widget_show_all(GTK_WIDGET(window));
+}
+
+static void callback_1()
+{
+    EXPECT_MESSAGE(zwlr_layer_surface_v1 .set_size 600 0);
+    EXPECT_MESSAGE(.create_buffer 600 1080); // size must match DEFAULT_OUTPUT_HEIGHT in common.h
+
+    gtk_layer_set_anchor(window, GTK_LAYER_SHELL_EDGE_LEFT, FALSE);
+}
+
+static void callback_2()
+{
+    EXPECT_MESSAGE(zwlr_layer_surface_v1 .set_size 600 700);
+    EXPECT_MESSAGE(.create_buffer 600 700);
+    gtk_layer_set_anchor(window, GTK_LAYER_SHELL_EDGE_BOTTOM, FALSE);
 }
 
 TEST_CALLBACKS(
     callback_0,
+    callback_1,
+    callback_2,
 )
