+++ test/integration-tests/test-anchors.c
@@ -9,27 +9,44 @@
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-#include "test-client-common.h"
+#include "integration-test-common.h"
 
 static GtkWindow* window;
 
-static void on_realize(GtkWidget *widget, gpointer _data)
+static void set_anchors(GtkWindow* window, gboolean top, gboolean bottom, gboolean left, gboolean right)
 {
-    (void)_data;
-    GdkWindow *gdk_window = gtk_widget_get_window(widget);
-    // This forces GTK to create a cairo surface for the window, which results in in being attached before our custom
-    // surface role objects are created/configured, which causes an error unless we clear the attached buffer
-    cairo_surface_t *cr = gdk_window_create_similar_surface (gdk_window, CAIRO_CONTENT_COLOR_ALPHA, 100, 100);
-    cairo_surface_destroy(cr);
+    gtk_layer_set_anchor(window, GTK_LAYER_SHELL_EDGE_TOP, top);
+    gtk_layer_set_anchor(window, GTK_LAYER_SHELL_EDGE_BOTTOM, bottom);
+    gtk_layer_set_anchor(window, GTK_LAYER_SHELL_EDGE_LEFT, left);
+    gtk_layer_set_anchor(window, GTK_LAYER_SHELL_EDGE_RIGHT, right);
 }
 
 static void callback_0()
 {
+    EXPECT_MESSAGE(zwlr_layer_surface_v1 .set_anchor 1);
+    EXPECT_MESSAGE(zwlr_layer_surface_v1 .set_anchor 2);
+    EXPECT_MESSAGE(zwlr_layer_surface_v1 .set_anchor 4);
+    EXPECT_MESSAGE(zwlr_layer_surface_v1 .set_anchor 8);
+    EXPECT_MESSAGE(zwlr_layer_surface_v1 .set_anchor 0);
+    EXPECT_MESSAGE(zwlr_layer_surface_v1 .set_anchor 3);
+    EXPECT_MESSAGE(zwlr_layer_surface_v1 .set_anchor 15);
+    EXPECT_MESSAGE(zwlr_layer_surface_v1 .set_anchor 0);
+
     window = create_default_window();
-    g_signal_connect(G_OBJECT(window), "realize", G_CALLBACK(on_realize), NULL);
     gtk_layer_init_for_window(window);
     gtk_widget_show_all(GTK_WIDGET(window));
+
+    set_anchors(window, TRUE, FALSE, FALSE, FALSE);
+    set_anchors(window, FALSE, TRUE, FALSE, FALSE);
+    set_anchors(window, FALSE, FALSE, TRUE, FALSE);
+    set_anchors(window, FALSE, FALSE, FALSE, TRUE);
+    set_anchors(window, FALSE, FALSE, FALSE, FALSE);
+    set_anchors(window, TRUE, TRUE, FALSE, FALSE);
+    set_anchors(window, TRUE, FALSE, TRUE, TRUE);
+    set_anchors(window, TRUE, TRUE, TRUE, TRUE);
+    set_anchors(window, FALSE, FALSE, FALSE, FALSE);
 }
+
 TEST_CALLBACKS(
     callback_0,
 )
