+++ test/integration-tests/test-get-keyboard-interactivity.c
@@ -9,7 +9,7 @@
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-#include "test-client-common.h"
+#include "integration-test-common.h"
 
 static GtkWindow* window;
 
@@ -17,21 +17,15 @@ static void callback_0()
 {
     window = create_default_window();
     gtk_layer_init_for_window(window);
-    gtk_layer_set_layer(window, GTK_LAYER_SHELL_LAYER_BOTTOM);
-    gtk_layer_set_namespace(window, "foobar");
-    gtk_layer_set_anchor(window, GTK_LAYER_SHELL_EDGE_BOTTOM, TRUE);
-    gtk_layer_set_exclusive_zone(window, 32);
-}
-
-static void callback_1()
-{
-    EXPECT_MESSAGE(zwlr_layer_shell_v1 .get_layer_surface 1 "foobar");
-    EXPECT_MESSAGE(zwlr_layer_surface_v1 .set_exclusive_zone 32);
-    EXPECT_MESSAGE(wl_surface .commit);
+    ASSERT_EQ(gtk_layer_get_keyboard_interactivity(window), FALSE, "%d");
+    gtk_layer_set_keyboard_interactivity(window, TRUE);
+    ASSERT_EQ(gtk_layer_get_keyboard_interactivity(window), TRUE, "%d");
     gtk_widget_show_all(GTK_WIDGET(window));
+    ASSERT_EQ(gtk_layer_get_keyboard_interactivity(window), TRUE, "%d");
+    gtk_layer_set_keyboard_interactivity(window, FALSE);
+    ASSERT_EQ(gtk_layer_get_keyboard_interactivity(window), FALSE, "%d");
 }
 
 TEST_CALLBACKS(
     callback_0,
-    callback_1,
 )
