+++ test/integration-tests/test-get-layer.c
@@ -9,31 +9,23 @@
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-#include "test-client-common.h"
+#include "integration-test-common.h"
 
 static GtkWindow* window;
 
 static void callback_0()
 {
-    // The mock server will automatically click on our window, triggering the menu to open
-
-    EXPECT_MESSAGE(zwlr_layer_shell_v1 .get_layer_surface);
-    EXPECT_MESSAGE(xdg_wm_base .get_xdg_surface);
-    EXPECT_MESSAGE(xdg_surface .get_popup);
-    EXPECT_MESSAGE(xdg_popup .grab);
-
-    window = GTK_WINDOW(gtk_window_new(GTK_WINDOW_TOPLEVEL));
-    GtkWidget *menu_bar = gtk_menu_bar_new();
-    gtk_container_add(GTK_CONTAINER(window), menu_bar);
-    GtkWidget *menu_item = gtk_menu_item_new_with_label("Popup menu");
-    gtk_container_add(GTK_CONTAINER(menu_bar), menu_item);
-    GtkWidget *submenu = gtk_menu_new();
-    gtk_menu_item_set_submenu(GTK_MENU_ITEM(menu_item), submenu);
-    GtkWidget *close_item = gtk_menu_item_new_with_label("Menu item");
-    gtk_menu_shell_append(GTK_MENU_SHELL(submenu), close_item);
-
+    window = create_default_window();
     gtk_layer_init_for_window(window);
+    gtk_layer_set_layer(window, GTK_LAYER_SHELL_LAYER_OVERLAY);
+    ASSERT_EQ(gtk_layer_get_layer(window), GTK_LAYER_SHELL_LAYER_OVERLAY, "%d");
     gtk_widget_show_all(GTK_WIDGET(window));
+    gtk_layer_set_layer(window, GTK_LAYER_SHELL_LAYER_TOP);
+    ASSERT_EQ(gtk_layer_get_layer(window), GTK_LAYER_SHELL_LAYER_TOP, "%d");
+    gtk_layer_set_layer(window, GTK_LAYER_SHELL_LAYER_BACKGROUND);
+    ASSERT_EQ(gtk_layer_get_layer(window), GTK_LAYER_SHELL_LAYER_BACKGROUND, "%d");
+    gtk_layer_set_layer(window, GTK_LAYER_SHELL_LAYER_BOTTOM);
+    ASSERT_EQ(gtk_layer_get_layer(window), GTK_LAYER_SHELL_LAYER_BOTTOM, "%d");
 }
 
 TEST_CALLBACKS(
