+++ test/integration-tests/test-get-namespace-default.c
@@ -9,7 +9,7 @@
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-#include "test-client-common.h"
+#include "integration-test-common.h"
 
 static GtkWindow* window;
 
@@ -17,16 +17,9 @@ static void callback_0()
 {
     window = create_default_window();
     gtk_layer_init_for_window(window);
-    gtk_layer_set_anchor(window, GTK_LAYER_SHELL_EDGE_BOTTOM, TRUE);
-    ASSERT_EQ(gtk_layer_get_exclusive_zone(window), 0, "%d");
-    gtk_layer_set_exclusive_zone(window, 12);
-    ASSERT_EQ(gtk_layer_get_exclusive_zone(window), 12, "%d");
     gtk_widget_show_all(GTK_WIDGET(window));
-    ASSERT_EQ(gtk_layer_get_exclusive_zone(window), 12, "%d");
-    gtk_layer_set_exclusive_zone(window, 0);
-    ASSERT_EQ(gtk_layer_get_exclusive_zone(window), 0, "%d");
-    gtk_layer_set_exclusive_zone(window, -1);
-    ASSERT_EQ(gtk_layer_get_exclusive_zone(window), -1, "%d");
+    const char *name_space = gtk_layer_get_namespace(window);
+    ASSERT_STR_EQ(name_space, "gtk-layer-shell");
 }
 
 TEST_CALLBACKS(
