+++ test/integration-tests/test-set-size-request.c
@@ -9,58 +9,58 @@
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-#include "test-client-common.h"
+#include "integration-test-common.h"
 
 static GtkWindow* window;
 
 static void callback_0()
 {
-    // First, anchor to bottom
-    EXPECT_MESSAGE(zwlr_layer_surface_v1 .set_exclusive_zone 200);
+    EXPECT_MESSAGE(zwlr_layer_surface_v1 .set_size 0 0);
 
     window = create_default_window();
-
     gtk_layer_init_for_window(window);
-    gtk_layer_auto_exclusive_zone_enable(window);
+    gtk_layer_set_layer(window, GTK_LAYER_SHELL_LAYER_BOTTOM);
+    gtk_layer_set_namespace(window, "foobar");
     gtk_layer_set_anchor(window, GTK_LAYER_SHELL_EDGE_BOTTOM, TRUE);
-    gtk_widget_set_size_request(GTK_WIDGET(window), 300, 200);
+    gtk_layer_set_anchor(window, GTK_LAYER_SHELL_EDGE_TOP, TRUE);
+    gtk_layer_set_anchor(window, GTK_LAYER_SHELL_EDGE_LEFT, TRUE);
+    gtk_layer_set_anchor(window, GTK_LAYER_SHELL_EDGE_RIGHT, TRUE);
     gtk_widget_show_all(GTK_WIDGET(window));
 }
 
 static void callback_1()
 {
-    // Resize
-    EXPECT_MESSAGE(zwlr_layer_surface_v1 .set_exclusive_zone 240);
+    EXPECT_MESSAGE(zwlr_layer_surface_v1 .set_size 600 700);
 
-    gtk_widget_set_size_request(GTK_WIDGET(window), 320, 240);
+    gtk_layer_set_anchor(window, GTK_LAYER_SHELL_EDGE_BOTTOM, FALSE);
+    gtk_layer_set_anchor(window, GTK_LAYER_SHELL_EDGE_TOP, FALSE);
+    gtk_layer_set_anchor(window, GTK_LAYER_SHELL_EDGE_LEFT, FALSE);
+    gtk_layer_set_anchor(window, GTK_LAYER_SHELL_EDGE_RIGHT, FALSE);
+    gtk_widget_set_size_request(GTK_WIDGET(window), 600, 700);
     gtk_window_resize(window, 1, 1);
 }
 
 static void callback_2()
 {
-    // Streatch across left edge
-    EXPECT_MESSAGE(zwlr_layer_surface_v1 .set_exclusive_zone 320);
+    EXPECT_MESSAGE(zwlr_layer_surface_v1 .set_size 0 700);
 
-    gtk_layer_set_anchor(window, GTK_LAYER_SHELL_EDGE_TOP, TRUE);
     gtk_layer_set_anchor(window, GTK_LAYER_SHELL_EDGE_LEFT, TRUE);
+    gtk_layer_set_anchor(window, GTK_LAYER_SHELL_EDGE_RIGHT, TRUE);
 }
 
 static void callback_3()
 {
-    // Stretch across top
-    EXPECT_MESSAGE(zwlr_layer_surface_v1 .set_exclusive_zone 240);
+    EXPECT_MESSAGE(zwlr_layer_surface_v1 .set_size 0 0);
 
-    gtk_layer_set_anchor(window, GTK_LAYER_SHELL_EDGE_BOTTOM, FALSE);
-    gtk_layer_set_anchor(window, GTK_LAYER_SHELL_EDGE_RIGHT, TRUE);
+    gtk_layer_set_anchor(window, GTK_LAYER_SHELL_EDGE_BOTTOM, TRUE);
+    gtk_layer_set_anchor(window, GTK_LAYER_SHELL_EDGE_TOP, TRUE);
 }
 
 static void callback_4()
 {
-    // Anchor right
-    EXPECT_MESSAGE(zwlr_layer_surface_v1 .set_exclusive_zone 320);
+    EXPECT_MESSAGE(zwlr_layer_surface_v1 .set_size 600 0);
 
-    gtk_layer_set_anchor(window, GTK_LAYER_SHELL_EDGE_TOP, FALSE);
-    gtk_layer_set_anchor(window, GTK_LAYER_SHELL_EDGE_LEFT, FALSE);
+    gtk_layer_set_anchor(window, GTK_LAYER_SHELL_EDGE_RIGHT, FALSE);
 }
 
 TEST_CALLBACKS(
