--- test/test-client-common/meson.build
@@ -1,2 +1,0 @@
-test_client_common_inc = include_directories('.')
-test_client_common_srcs = files('test-client-common.c')
